changed: Don't run custom script with --noimage changed: Remove /bin/sh header from custom script changed: Only report mounted partitions if --devices is specified changed: Create seperate fancify function to dump partition layout

+++ custom-backup-script.sh
@@ -1,21 +1,21 @@
-#!/bin/sh
-
-# Auto remove pagefile.sys & hiberfil.sys from NTFS Windows partitions
-IFS=' '
-for PART_NODEV in $BACKUP_PARTITIONS; do
-  TYPE=`blkid -s TYPE -o value "/dev/${PART_NODEV}"`
-  if [ "$TYPE" = "ntfs" ]; then
-    if mkdir -p /mnt/temp && ntfs-3g "/dev/${PART_NODEV}" /mnt/temp; then
-      if [ -f "/mnt/temp/hiberfil.sys" ]; then
-        echo "* Remove /dev/$PART_NODEV/hiberfil.sys"
-        rm /mnt/temp/hiberfil.sys
+if [ "$NO_IMAGE" != "1" ]; then
+  # Auto remove pagefile.sys & hiberfil.sys from NTFS Windows partitions
+  IFS=' '
+  for PART_NODEV in $BACKUP_PARTITIONS; do
+    TYPE=`blkid -s TYPE -o value "/dev/${PART_NODEV}"`
+    if [ "$TYPE" = "ntfs" ]; then
+      if mkdir -p /mnt/temp && ntfs-3g "/dev/${PART_NODEV}" /mnt/temp; then
+        if [ -f "/mnt/temp/hiberfil.sys" ]; then
+          echo "* Remove /dev/$PART_NODEV/hiberfil.sys"
+          rm /mnt/temp/hiberfil.sys
+        fi
+        if [ -f "/mnt/temp/pagefile.sys" ]; then
+          echo "* Remove /dev/$PART_NODEV/pagefile.sys"
+          rm /mnt/temp/pagefile.sys
+        fi
       fi
-      if [ -f "/mnt/temp/pagefile.sys" ]; then
-        echo "* Remove /dev/$PART_NODEV/pagefile.sys"
-        rm /mnt/temp/pagefile.sys
-      fi
+      
+      umount /mnt/temp
     fi
-    
-    umount /mnt/temp
-  fi
-done
+  done
+fi
